Split gametest
