+++ fabric-object-builder-api-v1/src/testmod/resources/fabric.mod.json
@@ -28,7 +28,7 @@
       "net.fabricmc.fabric.test.object.builder.VillagerTypeTest2"
     ],
     "fabric-gametest": [
-      "net.fabricmc.fabric.test.object.builder.BlockEntityTypeBuilderTest"
+      "net.fabricmc.fabric.test.object.builder.ObjectBuilderGameTest"
     ]
   }
 }
